+++ packages/react-schedules/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @incld/react-schedules
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @incld/client@0.2.1
+  - @incld/react@0.2.1
+
 ## 0.2.0
 
 ### Patch Changes