+++ packages/react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @incld/react
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @incld/client@0.2.1
+
 ## 0.2.0
 
 ### Minor Changes